--- a/aws_lambda_powertools/metrics/base.py
+++ b/aws_lambda_powertools/metrics/base.py
@@ -3,7 +3,6 @@ import json
 import logging
 import numbers
 import os
-import pathlib
 from collections import defaultdict
 from enum import Enum
 from typing import Any, Dict, List, Union
@@ -13,12 +12,9 @@ import fastjsonschema
 from ..shared import constants
 from ..shared.functions import resolve_env_var_choice
 from .exceptions import MetricUnitError, MetricValueError, SchemaValidationError
+from .schema import CLOUDWATCH_EMF_SCHEMA
 
 logger = logging.getLogger(__name__)
-
-_schema_path = pathlib.Path(__file__).parent / "./schema.json"
-with _schema_path.open() as f:
-    CLOUDWATCH_EMF_SCHEMA = json.load(f)
 
 MAX_METRICS = 100