--- a/aws_lambda_powertools/metrics/base.py
+++ b/aws_lambda_powertools/metrics/base.py
@@ -7,13 +7,13 @@ from collections import defaultdict
 from enum import Enum
 from typing import Any, Dict, List, Union
 
-import fastjsonschema
-
 from ..shared import constants
 from ..shared.functions import resolve_env_var_choice
+from ..shared.lazy_import import LazyLoader
 from .exceptions import MetricUnitError, MetricValueError, SchemaValidationError
 from .schema import CLOUDWATCH_EMF_SCHEMA
 
+fastjsonschema = LazyLoader("fastjsonschema", globals(), "fastjsonschema")
 logger = logging.getLogger(__name__)
 
 MAX_METRICS = 100