--- a/pytest.ini
+++ b/pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
 addopts = -ra --cov --cov-config=.coveragerc
 testpaths = ./tests
+markers =
+	perf: marks perf tests to be deselected (deselect with '-m "not perf"')