**Issue #, if available:** #294

## Description of changes:

<!--- One or two sentences as a summary of what's being changed -->

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply --> 

* [ ] [Meet tenets criteria](https://awslabs.github.io/aws-lambda-powertools-python/#tenets)
* [ ] Update tests
* [ ] Update docs
* [ ] PR title follows [conventional commit semantics](https://github.com/awslabs/aws-lambda-powertools-python/blob/376ec0a2ac0d2a40e0af5717bef42ff84ca0d1b9/.github/semantic.yml#L2)
* [ ] Create generic types for Tracer provider
* [ ] Create generic types for Tracer subsegments
* [ ] Convert EMF JSON Schema into native object to reduce I/O

## Notes

* **[Before] Logger import only** 0.33s user 0.15s system 82% cpu 0.582 total
* **[After] Logger import only** 0.07s user 0.02s system 97% cpu 0.102 total
* **[After] Tracer imported but not initialized** 0.08s user 0.02s system 84% cpu 0.115 total
* **[After] Tracer imported and initialized** 0.32s user 0.11s system 92% cpu 0.469 total
* **LazyLoader** adds roughly 10ms due to class instantiation and globals manipulation
* Hard lesson: Type annotations load lazily loaded library ahead of time diminishing its effectiveness

## Breaking change checklist

<!--- Ignore if it's not a breaking change -->

**RFC issue #**:

* [ ] Migration process documented
* [ ] Implement warnings (if it can live side by side)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.