@michaelbrewer 2021-03-01
@heitorlessa  nice, definitely handy for lambdas that just want structured logging or want to toggle tracing only for dev and test and disable on production.