@heitorlessa 2021-03-02
New flamegraph with these patches - Tracer no longer appears and removes ~400ms for all imports and customers once merged

![image](https://user-images.githubusercontent.com/3340292/109657655-05da3800-7b66-11eb-8f9c-f3a0eaa8b573.png)